ongoing...

+++ app/assets/scripts/app/user/controllers/userTable.js
@@ -1,27 +1,45 @@
 angular.module('user').controller('userTable', function($scope, userService) {
 
-    $scope.newUser = undefined;
+    $scope.newUser = {};
     $scope.isLoading = false;
 
     var refresh = function() {
         $scope.isLoading = true;
         userService.query(function(allUsers) {
             $scope.users = allUsers;
-            $scope.isLoading = false;
+            setTimeout(function(){
+                $scope.$apply(function(){$scope.isLoading = false;})
+            },1000);
+
         })
     }
 
     $scope.refreshUserTable = function(){refresh()}
 
     $scope.deleteUser = function(user) {
-        userService.remove({id: user.id},refresh)
+        if(window.confirm("are you sure ?")) {
+            userService.remove({id: user.id},refresh)
+        }
+    }
+
+    $scope.checkEmail = function(email) {
+        console.log(email)
+        return true;
     }
 
     $scope.createUser = function(user) {
-        userService.save(user,refresh)
+        userService.save(user,
+            function(){
+                $scope.newUser = {};
+                refresh();
+            },
+            function(response){
+                console.log(response.data)
+            }
+        )
     }
 
-    $scope.updateUser = function(user) {
+    $scope.update = function(user) {
         userService.update({id: user.id}, {name: user.name, email: user.email}, refresh)
     }
